Rename PortType.REAL to PortType.SERIAL
REAL was named for the distinction it originally carried -- a real serial port
versus a fake one -- from back when those were the only two kinds. Alongside
SOCKET, BLUETOOTH and RTT it says nothing about the medium, and now that it is
the transport's `kind` it reads badly at every use site:
`transport.kind === PortType.REAL` for the flow-control check being the worst.

Pure rename: PortType is not persisted anywhere, so no migration.

Also documents why PortType and ConnectionType exist side by side as nearly
the same set. They differ in one place that matters: ConnectionType.SERIAL_PORT
covers a real port and a fake-port generator, which map to different transports
here. Collapsing them would need an app-data migration since ConnectionType is
persisted, so the duplication is recorded rather than removed.

+++ src/renderer/src/model/ConnController/ConnController.characterisation.spec.ts
@@ -111,7 +111,7 @@ describe('ConnController connection lifecycle', () => {
   async function openSerial() {
     app.settings.portConfiguration.setConnectionType(ConnectionType.SERIAL_PORT);
     app.settings.portConfiguration.selectedSerialPort = { path: 'COM3' };
-    conn.lastSelectedPortType = PortType.REAL;
+    conn.lastSelectedPortType = PortType.SERIAL;
     return conn.openConnection();
   }
 
@@ -384,7 +384,7 @@ describe('ConnController connection lifecycle', () => {
       app.settings.portConfiguration.selectedSerialPort = { path: 'COM1' };
 
       conn.setSelectedPort({ path: 'COM7' } as any);
-      conn.lastSelectedPortType = PortType.REAL;
+      conn.lastSelectedPortType = PortType.SERIAL;
       await conn.openConnection();
 
       const openedWith = (window.electronAPI.serial.openPort as unknown as Mock).mock.calls[0][0];
